# ndc-aggregator/server/docs/swagger.yaml
openapi: 3.0.3
info:
  title: IATA Digital Identity B2C PoC — NDC Aggregator API
  description: |
    REST API for the NDC Aggregator Proof-of-Concept.
    Integrates IATA Digital Identity (Verifiable Credentials via OpenID4VP)
    with multi-airline NDC capabilities (AirShopping, OfferPrice, OrderCreate).

    **Authentication** is handled via two mutually exclusive header patterns:
    - `X-Agency-Token` — session token from agency login (UC1), triggers server-side VP generation
    - `X-Verification-Id` — session ID from QR-based verification (UC2)
  version: 1.0.0
  contact:
    name: Dreamix
servers:
  - url: http://localhost:3000/api
    description: Local development
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Agency login
      description: Validates agency credentials and returns a session token for UC1.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    format: uuid
                    description: Agency session token (use as X-Agency-Token header)
        '401':
          description: Wrong credentials
  /auth/vc:
    post:
      tags: [Auth]
      summary: Upload agency Verifiable Credential
      description: Stores a W3C VCDM Verifiable Credential for server-side VP generation (UC1).
      security:
        - AgencyToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifiableCredential'
      responses:
        '200':
          description: VC stored successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  vc:
                    $ref: '#/components/schemas/VCInfo'
        '400':
          description: Invalid VC format
        '401':
          description: Invalid or expired agency token

    get:
      tags: [Auth]
      summary: Get stored VC status
      security:
        - AgencyToken: []
      responses:
        '200':
          description: VC info
          content:
            application/json:
              schema:
                type: object
                properties:
                  hasVC:
                    type: boolean
                  vc:
                    $ref: '#/components/schemas/VCInfo'
        '401':
          description: Invalid or expired agency token

    delete:
      tags: [Auth]
      summary: Delete stored agency VC
      security:
        - AgencyToken: []
      responses:
        '200':
          description: Deletion result
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: boolean
                  message:
                    type: string
        '401':
          description: Invalid or expired agency token

  # ── Credentials ─────────────────────────────────────────
...
components:
  securitySchemes:
    AgencyToken:
      type: apiKey
      in: header
      name: X-Agency-Token
      description: Session token from POST /auth/login (UC1)
    VerificationId:
      type: apiKey
      in: header
      name: X-Verification-Id
      description: Session ID from QR verification (UC2)
  schemas:
...
    VerifiableCredential:
      type: object
      description: W3C VCDM Verifiable Credential
      required: [type, credentialSubject]
      properties:
        type:
          type: array
          items:
            type: string
        credentialSubject:
          type: object
          required: [iata_number]
          properties:
            iata_number:
              type: string
    VCInfo:
      type: object
      nullable: true
      properties:
        hasVC:
          type: boolean
        iataNumber:
          type: string
        uploadedAt:
          type: string
          format: date-time
...